Move TsIgnore to .Typescript namespace
